\name{RKTest-class}
\alias{RKTest
-class}
\alias{RKTest}
\title{S4 class RKTest}
\description{This class is used internally by \code{\link[rkwardtests:rktest.makeplugintests]{rktest.makeplugintests}}.}
\keyword{classes}
\author{Thomas Friedrichsmeier \email{thomas.friedrichsmeier@ruhr-uni-bochum.de}}
\section{Slots}{\describe{\item{\code{id}:}{}\item{\code{call}:}{}\item{\code{fuzzy_output}:}{}\item{\code{expect_error}:}{}\item{\code{libraries}:}{}\item{\code{files}:}{}}}

