% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rktest.getTempDir.R
\name{rktest.getTempDir}
\alias{rktest.getTempDir}
\title{Get path to the temporary directory}
\usage{
rktest.getTempDir()
}
\value{
Either a character string, or FALSE.
}
\description{
This function will return either the local path to the temporary directory where
all test results have been saved to, or FALSE if none exitsts.
}
\details{
Get the path to the recent temporary directory, if one exists.
}
\examples{
rktest.getTempDir()
}
\seealso{
\code{\link[rkwardtests:rktest.makeplugintests]{rktest.makeplugintests}}
}
\author{
Meik Michalke \email{meik.michalke@uni-duesseldorf.de}
}
\keyword{utilities}
