% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RKTest-class.R
\docType{class}
\name{RKTest}
\alias{RKTest}
\title{S4 class RKTest}
\description{
This class is used internally by \code{\link[rkwardtests:rktest.makeplugintests]{rktest.makeplugintests}}.
}
\section{Slots}{

\describe{
\item{\code{id}}{A unique character string}

\item{\code{call}}{A function to be called}

\item{\code{fuzzy_output}}{Allow fuzzy results}

\item{\code{expect_error}}{Expect errors}

\item{\code{libraries}}{A charcter vector naming needed libraries}

\item{\code{files}}{A character vector naming needed files, path relative to the test standards directory}

\item{\code{record.all.commands}}{Should synchronization commands and commands to generate run-again-links be included in the command recording? Generally, this should be FALSE (the default).}

\item{\code{ignore}}{May include one of more of "output", "messages", "commands", for skipping comparison of these against the standard, completely.}
}}

\author{
Thomas Friedrichsmeier \email{thomas.friedrichsmeier@ruhr-uni-bochum.de}
}
\keyword{classes}
