% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rktest.replaceRunAgainLink.R
\name{rktest.replaceRunAgainLink}
\alias{rktest.replaceRunAgainLink}
\alias{.rk.rerun.plugin.link.replacement}
\title{Replace "Run again" link in RKWard}
\usage{
rktest.replaceRunAgainLink(restore=FALSE)
}
\arguments{
\item{restore}{Logical: If TRUE, restore the original behaviour.}
}
\value{
Replaces the "Run again" link in RKWard with the code that would have been called, or vice versa.
}
\description{
You can use this to temporarily replace .rk.rerun.plugin.link (see example below).
This way, after running a plugin, you are shown the call needed to run this
plugin with those settings, instead of the link.

This code can be used in a plugin test suite.
}
\details{
Replace "Run again" link in RKWard with code
}
\examples{
rktest.replaceRunAgainLink()
}
\seealso{
\code{\link[rkwardtests:RKTestSuite]{RKTestSuite-class}}, \code{\link[rkwardtests:rktest.makeplugintests]{rktest.makeplugintests}}
}
\author{
Thomas Friedrichsmeier \email{thomas.friedrichsmeier@ruhr-uni-bochum.de}, Meik Michalke \email{meik.michalke@uni-duesseldorf.de}
}
\keyword{utilities}
