\name{RKTestResult-class}
\alias{RKTestResult
-class}
\alias{RKTestResult}
\title{S4 class RKTestResult}
\description{Class RKTestResult}
\details{This class is used internally by \code{\link[rkwardtests:rktest.makeplugintests]{rktest.makeplugintests}}.}
\keyword{classes}
\author{Thomas Friedrichsmeier \email{thomas.friedrichsmeier@ruhr-uni-bochum.de}}
\section{Slots}{\describe{\item{\code{id}:}{}\item{\code{code_match}:}{}\item{\code{output_match}:}{}\item{\code{message_match}:}{}\item{\code{error}:}{}\item{\code{passed}:}{}}}

