% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RKTestResult-class.R
\docType{class}
\name{RKTestResult}
\alias{RKTestResult}
\title{S4 Class RKTestResult}
\description{
This class is used internally by \code{\link[rkwardtests:rktest.makeplugintests]{rktest.makeplugintests}}.
}
\section{Slots}{

\describe{
\item{\code{id}}{A unique character string naming a test.}

\item{\code{code_match}}{A character string indicating whether the run code matched the standard.}

\item{\code{output_match}}{A character string indicating whether the resulting output matched the standard.}

\item{\code{message_match}}{A character string indicating whether the resulting R messages matched the standard.}

\item{\code{error}}{A character string indicating errors.}

\item{\code{missing_libs}}{A character string indicating missing libraries.}

\item{\code{passed}}{Logical: Did the test pass?}
}}

\author{
Thomas Friedrichsmeier \email{thomas.friedrichsmeier@ruhr-uni-bochum.de}
}
\keyword{classes}
