\name{RKTestSuite-class}
\alias{RKTestSuite
-class}
\alias{RKTestSuite}
\title{S4 class RKTestSuite}
\description{Class RKTestSuite}
\details{This class is used to create test suite objects that can be fed to \code{\link[rkwardtests:rktest.makeplugintests]{rktest.makeplugintests}}.}
\keyword{classes}
\author{Thomas Friedrichsmeier \email{thomas.friedrichsmeier@ruhr-uni-bochum.de}}
\section{Slots}{\describe{\item{\code{id}:}{}\item{\code{libraries}:}{}\item{\code{initCalls}:}{}\item{\code{tests}:}{}\item{\code{postCalls}:}{}}}

