% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RKTestSuite-class.R
\docType{class}
\name{RKTestSuite}
\alias{RKTestSuite}
\title{S4 Class RKTestSuite}
\description{
This class is used to create test suite objects that can be fed to \code{\link[rkwardtests:rktest.makeplugintests]{rktest.makeplugintests}}.
}
\section{Slots}{

\describe{
\item{\code{id}}{A unique character string to identify a test suite}

\item{\code{libraries}}{A charcter vector naming libraries that the test suite depends on.}

\item{\code{initCalls}}{A list of functions to be run before any tests, e.g. to load libraries or data objects.}

\item{\code{tests}}{A list of the actual plugin tests.}

\item{\code{postCalls}}{A list of functions to be run after all tests, e.g. to clean up.}
}}

\author{
Thomas Friedrichsmeier \email{thomas.friedrichsmeier@ruhr-uni-bochum.de}
}
\keyword{classes}
