% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkwardtests-package.R
\docType{package}
\name{rkwardtests-package}
\alias{rkwardtests-package}
\alias{rkwardtests}
\title{RKWard Plugin Test Suite Framework.}
\description{
A set of functions, classes and methods to test plugins that were written for RKWard.
}
\details{
RKWard Plugin Test Suite Framework.

\tabular{ll}{
Package: \tab rkwardtests\cr
Type: \tab Package\cr
Version: \tab 0.5.7\cr
Date: \tab 2012-06-03\cr
Depends: \tab R (>= 2.9.0),methods\cr
Encoding: \tab UTF-8\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
URL: \tab https://rkward.kde.org\cr
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rkward.kde.org}
}

}
\author{
Thomas Friedrichsmeier, Meik Michalke
}
\keyword{package}
