\docType{package}
\name{rkwardtests-package}
\alias{rkwardtests}
\alias{rkwardtests-package}
\title{RKWard Plugin Test Suite Framework.}
\description{
  RKWard Plugin Test Suite Framework.
}
\details{
  \tabular{ll}{ Package: \tab rkwardtests\cr Type: \tab
  Package\cr Version: \tab 0.5.7\cr Date: \tab
  2012-06-03\cr Depends: \tab R (>= 2.9.0),methods\cr
  Encoding: \tab UTF-8\cr License: \tab GPL (>= 2)\cr
  LazyLoad: \tab yes\cr URL: \tab
  http://rkward.sourceforge.net\cr }

  A set of functions, classes and methods to test plugins
  that were written for RKWard.
}
\author{
  Thomas Friedrichsmeier, Meik Michalke
}
\keyword{package}

