\name{rkwardtests-package}
\alias{rkwardtests-package}
\alias{rkwardtests-package}
\title{RKWard Plugin Test Suite Framework.}
\description{RKWard Plugin Test Suite Framework.}
\details{\tabular{ll}{
Package: \tab rkwardtests\cr
Type: \tab Package\cr
Version: \tab 0.5.5\cr
Date: \tab 2010-10-14\cr
Depends: \tab R (>= 2.9.0),methods\cr
Encoding: \tab UTF-8\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
URL: \tab http://rkward.sourceforge.net \cr
}

A set of functions, classes and methods to test plugins that were written for RKWard.}
\alias{rkwardtests-package}
\alias{rkwardtests}
\docType{package}
\author{Thomas Friedrichsmeier \email{thomas.friedrichsmeier@ruhr-uni-bochum.de}, Meik Michalke \email{meik.michalke@uni-duesseldorf.de}}
\keyword{package}
\seealso{\code{\link[rkwardtests:rktest.makeplugintests]{rktest.makeplugintests}}}

