\name{show}
\alias{show}
\alias{show,RKTestResult-method}
\title{show method for objects of class RKTestResult}
\arguments{
  \item{object}{An object of class RKTestResult}
}
\description{
  show method for S4 objects of class RKTestResult
}
\details{
  Prints a summary of plugin test results.
}
\examples{
\dontrun{
rktest.makeplugintests("rkward_application_tests.R")
}
}
\author{
  Thomas Friedrichsmeier
  \email{thomas.friedrichsmeier@ruhr-uni-bochum.de}
}
\keyword{methods}

