\name{show}
\alias{show,-method}
\alias{show}
\title{show method for objects of class RKTestResult}
\usage{\method{show}{RKTestResult}(object)
}
\description{show method for S4 objects of class RKTestResult}
\details{Prints a summary of plugin test results.}
\alias{show,RKTestResult-method}
\author{Thomas Friedrichsmeier \email{thomas.friedrichsmeier@ruhr-uni-bochum.de}}
\keyword{methods}
\arguments{\item{object}{An object of class RKTestResult}
}
\examples{\dontrun{
rktest.makeplugintests("rkward_application_tests.R")
}}

